+++ README.md
@@ -1,11 +1,7 @@
 # `YIFANCORE`
 ——作者:YifanZhuZhu
 
-## ** 更新: **
-```
-power函数采用了新的算法更快了
-鸣谢:TakeOff0518
-```
+
 ### `注意事项`
 ```
 .h用于c语言文件,.hpp用于c++文件
