+++ test_power.cpp
@@ -1,0 +1,10 @@
+#include<cstdio>
+#include"yifancore.hpp"
+int main()
+{
+	int m,n;
+	scanf("%d%d",&m,&n);
+	long long a=power(m,n);
+	printf("%lld",a);
+	return 0;
+}
