+++ test_split.cpp
@@ -1,0 +1,10 @@
+#include<cstdio>
+#include"yifancore.hpp"
+int main()
+{
+	int a,m,n;
+	scanf("%d%d",&m,&n);
+	a=split(m,n);
+	printf("%d",a);
+	return 0;
+}
